Update danielpratt_gpt_understanding.cs
Asked chatgpt how the portion in FixedUpdate() works and if there was a better way to implement this (https://chatgpt.com/share/66fe260d-4bc4-800f-b949-9e1506b73be4).

There was! So now it is much easier to read and I added some comments explaining what that section actually did.